+++ src/plugin/radarscreen/RadarScreenFactory.cpp
@@ -1,4 +1,5 @@
 #include "ConfigurableDisplayCollection.h"
+#include "MenuToggleableDisplayFactory.h"
 #include "PositionResetCommand.h"
 #include "RadarScreenFactory.h"
 #include "ScreenControlsBootstrap.h"
@@ -22,6 +23,7 @@
 #include "releases/ReleaseModule.h"
 #include "sectorfile/SectorFileBootstrap.h"
 #include "srd/SrdModule.h"
+#include "wake/WakeModule.h"
 
 using UKControllerPlugin::Api::BootstrapConfigurationMenuItem;
 using UKControllerPlugin::Bootstrap::HelperBootstrap;
@@ -54,6 +56,7 @@ namespace UKControllerPlugin::RadarScreen {
         AsrEventHandlerCollection userSettingHandlers;
         ConfigurableDisplayCollection configurableDisplays;
         CommandHandlerCollection commandHandlers;
+        MenuToggleableDisplayFactory displayFactory(*this->persistence.pluginFunctionHandlers, configurableDisplays);
 
         // Run bootstrap
         BootstrapConfigurationMenuItem(persistence, configurableDisplays);
@@ -109,6 +112,7 @@ namespace UKControllerPlugin::RadarScreen {
         PrenoteModule::BootstrapRadarScreen(this->persistence, renderers);
         Departure::BootstrapRadarScreen(this->persistence, renderers, configurableDisplays, userSettingHandlers);
         MissedApproach::BootstrapRadarScreen(this->persistence, renderers, configurableDisplays, userSettingHandlers);
+        Wake::BootstrapRadarScreen(this->persistence, renderers, userSettingHandlers, displayFactory);
 
         // Register command for position resets
         this->persistence.commandHandlers->RegisterHandler(std::make_shared<PositionResetCommand>(renderers));