+++ test/plugin/hold/HoldDisplayFactoryTest.cpp
@@ -6,8 +6,9 @@
 #include "hold/HoldDisplayFactory.h"
 #include "hold/HoldManager.h"
 #include "hold/HoldingData.h"
-#include "navaids/NavaidCollection.h"
 #include "hold/PublishedHoldCollection.h"
+#include "list/PopupListFactory.h"
+#include "navaids/NavaidCollection.h"
 #include "navaids/Navaid.h"
 #include "plugin/FunctionCallEventHandler.h"
 
@@ -20,6 +21,7 @@ using UKControllerPlugin::Hold::HoldDisplayFactory;
 using UKControllerPlugin::Hold::HoldingData;
 using UKControllerPlugin::Hold::HoldManager;
 using UKControllerPlugin::Hold::PublishedHoldCollection;
+using UKControllerPlugin::List::PopupListFactory;
 using UKControllerPlugin::Navaids::Navaid;
 using UKControllerPlugin::Navaids::NavaidCollection;
 using UKControllerPlugin::Plugin::FunctionCallEventHandler;
@@ -34,8 +36,9 @@ namespace UKControllerPluginTest::Hold {
     {
         public:
         HoldDisplayFactoryTest()
-            : dialogManager(dialogProvider), callsignSelectionFactory(functionHandlers, mockPlugin),
-              navaid({1, "TIMBA", EuroScopePlugIn::CPosition()}), holdManager(mockApi, taskRunner),
+            : dialogManager(dialogProvider), popupFactory(functionHandlers, mockPlugin),
+              callsignSelectionFactory(popupFactory), navaid({1, "TIMBA", EuroScopePlugIn::CPosition()}),
+              holdManager(mockApi, taskRunner),
               factory(mockPlugin, holdManager, navaids, holds, dialogManager, callsignSelectionFactory)
 
         {
@@ -49,6 +52,7 @@ namespace UKControllerPluginTest::Hold {
         DialogManager dialogManager;
         NiceMock<MockEuroscopePluginLoopbackInterface> mockPlugin;
         FunctionCallEventHandler functionHandlers;
+        PopupListFactory popupFactory;
         CallsignSelectionListFactory callsignSelectionFactory;
         HoldingData holdData = {1, "TIMBA", "TIMBA TEST", 7000, 15000, 360, "left", {}};
         PublishedHoldCollection holds;